REFACTOR: 내부 호출 컨트롤러 분리
