+++ src/main/java/com/moyeorak/auth_service/config/SecurityConfig.java
@@ -39,7 +39,7 @@ public class SecurityConfig {
 
                         // --- 공개 엔드포인트 ---
                         .requestMatchers(
-                                "/api/users/**",
+                                "/internal/users/**",
                                 "/api/users/signup",
                                 "/api/auth/login",
                                 "/swagger-ui/**",
